requirements.txt
torch==1.6.0
espnet==0.9.2
parallel_wavegan==0.4.6
deepspeech
pyaudio
pydub
